More refactoring. Added ElixirScript.Macro.Env module
