+++ lib/elixir_script/universal/macro/env.ex
@@ -1,0 +1,19 @@
+defmodule ElixirScript.Macro.Env do
+  defstruct [
+    module: nil,
+    file: nil,
+    line: 0,
+    function: nil,
+    context: nil,
+    aliases: [],
+    requires: [],
+    functions: [],
+    macros: [],
+    macro_aliases: [],
+    context_modules: [],
+    vars: [],
+    export_vars: nil,
+    lexical_tracker: nil,
+    caller: nil
+  ]
+end
